requirements.txt
gekko
attrs
argparse
pyyaml
